requirements.txt
requests~=2.31.0
configparser~=6.0.0
PyYAML~=6.0.1
pytest~=7.4.2